SERVER-23624 Allow wiredTigerCacheSizeGB to be specified as a float
GitOrigin-RevId: f76495aff7de521fce4ef50c1766a7aca85808fc

+++ src/mongo/db/modules/enterprise/src/inmemory/inmemory_global_options.h
@@ -42,7 +42,7 @@ public:
     Status add(moe::OptionSection* options);
     Status store(const moe::Environment& params, const std::vector<std::string>& args);
 
-    size_t inMemorySizeGB;
+    double inMemorySizeGB;
     size_t statisticsLogDelaySecs;
     std::string engineConfig;
